requirements.txt
sentence-transformers

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu101
torchvision==0.8.2+cu101
torchaudio===0.7.2
